Always set attribute_will_change! when position is set
This prevents problems where the scope changes but the position remains the same as what's in the database but it was explicitly set again (indicating that the user wants to have the item at that position). If the scope changes and the position isn't explicitly set, we reset it to be `nil` and move the item to the end of the scope.