Fix healer when model uses an ordering default scope
When using an ordering default scope in the model, the query generated by the
healer with `model.select(*positioning_columns).distinct` is invalid:

    ERROR:  for SELECT DISTINCT, ORDER BY expressions must appear in select
    list (PG::InvalidColumnReference)

With a default scope that orders the records, Rails automatically adds the
default scope to all queries, including the SELECT DISTINCT above. This breaks
the healer, because it does not add the default ordering scope to the select
list. Unscoping the query before selecting removes the default query and makes
the healer work. This should not impact other functionality, since any ordering
required when healing must be passed explicitely anyway.